Refactor CreateIssue
