--- src/components/Issue/CreateIssue.css
@@ -1,19 +1,0 @@
-.input-title {
-  color: black;
-  font-size: 20px;
-  margin-top: 30px;
-}
-
-.select-label {
-  margin-top: 30px;
-  color: black;
-  font-size: 20px;
-  margin-bottom: 10px;
-}
-
-.dialog-title .MuiTypography-h6 {
-  font-size: 30px;
-  font-weight: bolder;
-  margin-bottom: 15px;
-  text-align: center;
-}
